pushing updated Actions classes
BaseAction, SimulateStep, AddSettlement, AddFacility Are ready.
AddPlan - all function are written but there is one gap, please check line 47 i marked it as a comment not to pass you an error.
i didnt know what initial scores to enter to the balanceSelection
(i mentioned the issue in the whatsapp voice messege)

Co-Authored-By: arikshvarts <104718686+arikshvarts@users.noreply.github.com>